Fixed mouse input scaling issue when back buffer size and window client size are not equal.
Note: there seems to be a bug in Mono where GameWindow.ClientBounds does not seem to be updated on game startup, only when you move the window will it give out correct values. It seems to use the window size at first and not the actual window client size (without the titlebar and all). This bug will only show on the same cases as the previous bug.